requirements.txt
# demo requirements.txt
python-magic
spdx_matcher
google-cloud-storage